+++ src/main/java/com/yahoo/bullet/querying/Querier.java
@@ -16,6 +16,11 @@ import com.yahoo.bullet.parsing.Projection;
 import com.yahoo.bullet.parsing.Query;
 import com.yahoo.bullet.parsing.Window;
 import com.yahoo.bullet.postaggregations.PostStrategy;
+import com.yahoo.bullet.querying.operations.AggregationOperations;
+import com.yahoo.bullet.querying.operations.FilterOperations;
+import com.yahoo.bullet.querying.operations.PostAggregationOperations;
+import com.yahoo.bullet.querying.operations.ProjectionOperations;
+import com.yahoo.bullet.querying.operations.WindowingOperations;
 import com.yahoo.bullet.record.BulletRecord;
 import com.yahoo.bullet.record.BulletRecordProvider;
 import com.yahoo.bullet.result.Clip;
