+++ src/test/java/com/yahoo/bullet/querying/operations/ProjectionOperationsTest.java
@@ -3,7 +3,7 @@
  *  Licensed under the terms of the Apache License, Version 2.0.
  *  See the LICENSE file associated with the project for terms.
  */
-package com.yahoo.bullet.querying;
+package com.yahoo.bullet.querying.operations;
 
 import com.yahoo.bullet.common.BulletConfig;
 import com.yahoo.bullet.parsing.Projection;
